--- resources/views/demande__salles/create.blade.php
@@ -1,50 +1,0 @@
-@extends('layouts.app')
-
-@section('content')
-
-    <div class="panel panel-default">
-
-        <div class="panel-heading clearfix">
-            
-            <span class="pull-left">
-                <h4 class="mt-5 mb-5">Create New Demande  Salle</h4>
-            </span>
-
-            <div class="btn-group btn-group-sm pull-right" role="group">
-                <a href="{{ route('demande__salles.demande__salle.index') }}" class="btn btn-primary" title="Show All Demande  Salle">
-                    <span class="glyphicon glyphicon-th-list" aria-hidden="true"></span>
-                </a>
-            </div>
-
-        </div>
-
-        <div class="panel-body">
-        
-            @if ($errors->any())
-                <ul class="alert alert-danger">
-                    @foreach ($errors->all() as $error)
-                        <li>{{ $error }}</li>
-                    @endforeach
-                </ul>
-            @endif
-
-            <form method="POST" action="{{ route('demande__salles.demande__salle.store') }}" accept-charset="UTF-8" id="create_demande__salle_form" name="create_demande__salle_form" class="form-horizontal">
-            {{ csrf_field() }}
-            @include ('demande__salles.form', [
-                                        'demandeSalle' => null,
-                                      ])
-
-                <div class="form-group">
-                    <div class="col-md-offset-2 col-md-10">
-                        <input class="btn btn-primary" type="submit" value="Add">
-                    </div>
-                </div>
-
-            </form>
-
-        </div>
-    </div>
-
-@endsection
-
-
