--- resources/views/demande__salles/edit.blade.php
@@ -1,52 +1,0 @@
-@extends('layouts.app')
-
-@section('content')
-
-    <div class="panel panel-default">
-  
-        <div class="panel-heading clearfix">
-
-            <div class="pull-left">
-                <h4 class="mt-5 mb-5">{{ !empty($title) ? $title : 'Demande  Salle' }}</h4>
-            </div>
-            <div class="btn-group btn-group-sm pull-right" role="group">
-
-                <a href="{{ route('demande__salles.demande__salle.index') }}" class="btn btn-primary" title="Show All Demande  Salle">
-                    <span class="glyphicon glyphicon-th-list" aria-hidden="true"></span>
-                </a>
-
-                <a href="{{ route('demande__salles.demande__salle.create') }}" class="btn btn-success" title="Create New Demande  Salle">
-                    <span class="glyphicon glyphicon-plus" aria-hidden="true"></span>
-                </a>
-
-            </div>
-        </div>
-
-        <div class="panel-body">
-
-            @if ($errors->any())
-                <ul class="alert alert-danger">
-                    @foreach ($errors->all() as $error)
-                        <li>{{ $error }}</li>
-                    @endforeach
-                </ul>
-            @endif
-
-            <form method="POST" action="{{ route('demande__salles.demande__salle.update', $demandeSalle->id) }}" id="edit_demande__salle_form" name="edit_demande__salle_form" accept-charset="UTF-8" class="form-horizontal">
-            {{ csrf_field() }}
-            <input name="_method" type="hidden" value="PUT">
-            @include ('demande__salles.form', [
-                                        'demandeSalle' => $demandeSalle,
-                                      ])
-
-                <div class="form-group">
-                    <div class="col-md-offset-2 col-md-10">
-                        <input class="btn btn-primary" type="submit" value="Update">
-                    </div>
-                </div>
-            </form>
-
-        </div>
-    </div>
-
-@endsection
